requirements.txt
PyQt5==5.15.9
keyboard==0.13.5
requests==2.31.0
Pillow==10.1.0 